+++ execution/market_snapshot.go
@@ -153,10 +153,7 @@ func (m *Market) changed() bool {
 		m.as.Changed() ||
 		m.peggedOrders.Changed() ||
 		m.expiringOrders.Changed() ||
-		m.equityShares.Changed() ||
-		m.liquidity.Changed() ||
-		m.position.Changed() ||
-		m.tsCalc.Changed())
+		m.equityShares.Changed())
 }
 
 func (m *Market) getState() *types.ExecMarket {