+++ snapshot/engine_test.go
@@ -13,10 +13,12 @@ import (
 	"code.vegaprotocol.io/vega/types"
 	tmocks "code.vegaprotocol.io/vega/types/mocks"
 
+	"github.com/cosmos/iavl"
 	"github.com/golang/mock/gomock"
 	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	db "github.com/tendermint/tm-db"
 )
 
 type tstEngine struct {
@@ -48,6 +50,56 @@ func getTestEngine(t *testing.T) *tstEngine {
 	}
 }
 
+// returns an avl tree populated with some payloads.
+func getPopulatedTree(t *testing.T) *iavl.MutableTree {
+	t.Helper()
+	testPayloads := []types.Payload{
+		{
+			Data: &types.PayloadAppState{
+				AppState: &types.AppState{
+					Height: 64,
+				},
+			},
+		},
+		{
+			Data: &types.PayloadGovernanceActive{
+				GovernanceActive: &types.GovernanceActive{},
+			},
+		},
+		{
+			Data: &types.PayloadGovernanceEnacted{
+				GovernanceEnacted: &types.GovernanceEnacted{},
+			},
+		},
+		{
+			Data: &types.PayloadDelegationActive{
+				DelegationActive: &types.DelegationActive{},
+			},
+		},
+		{
+			Data: &types.PayloadEpoch{
+				EpochState: &types.EpochState{
+					Seq:                  7,
+					ReadyToStartNewEpoch: true,
+				},
+			},
+		},
+	}
+
+	tree, err := iavl.NewMutableTree(db.NewMemDB(), 0)
+	tree.Load()
+	require.NoError(t, err)
+
+	for _, p := range testPayloads {
+		v, _ := proto.Marshal(p.IntoProto())
+		tree.Set([]byte(p.GetTreeKey()), v)
+	}
+
+	// Save it
+	tree.SaveVersion()
+	return tree
+}
+
 // basic engine functionality tests.
 func TestEngine(t *testing.T) {
 	t.Run("Adding a provider calls what we expect on the state provider", testAddProviders)
@@ -60,6 +112,43 @@ func TestRestore(t *testing.T) {
 	t.Run("Restoring a snapshot from chain works as expected", testReloadSnapshot)
 	t.Run("Restoring replay protectors replaces the provider as expected", testReloadReplayProtectors)
 	t.Run("Restoring a snapshot calls the post-restore callback if available", testReloadRestore)
+}
+
+func TestTreeToSnapshot(t *testing.T) {
+	t.Run("A tree can be exported serialised and then imported", testTreeExportImport)
+}
+
+func testTreeExportImport(t *testing.T) {
+	// get a avl tree with some payloads in it
+	tree := getPopulatedTree(t)
+
+	// export the tree into snapshot data
+	snap, err := types.SnapshotFromTree(tree.ImmutableTree)
+	require.NoError(t, err)
+	require.Equal(t, snap.Hash, tree.Hash())
+	require.Equal(t, snap.Meta.Version, tree.Version())
+	require.Equal(t, len(snap.Nodes), int(tree.Size()))
+	// We expect more nodehashes than nodes since nodes only contain the leaf nodes
+	// with payloads whereas nodehashes contain the payload-less subtree-roots
+	require.Greater(t, len(snap.Meta.NodeHashes), len(snap.Nodes))
+
+	// Note IRL it would be now that snapshot is serialised and sent
+	// via TM to the node restoring from a snapshot
+
+	// Make a new tree waiting to import the snapshot
+	importedTree, err := iavl.NewMutableTree(db.NewMemDB(), 0)
+	importedTree.Load()
+	require.NoError(t, err)
+
+	// import the snapshot data into a new avl tree
+	err = snap.TreeFromSnapshot(importedTree)
+	require.NoError(t, err)
+
+	// The new tree should be identical to the previous
+	assert.Equal(t, tree.Hash(), importedTree.Hash())
+	assert.Equal(t, tree.Size(), importedTree.Size())
+	assert.Equal(t, tree.Height(), importedTree.Height())
+	assert.Equal(t, tree.Version(), importedTree.Version())
 }
 
 func testAddProviders(t *testing.T) {